% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{chartOutput}
\alias{chartOutput}
\title{Chart Output}
\usage{
chartOutput(outputID, width = "100\%", height = "400px")
}
\arguments{
\item{outputID}{Output variable from which to read the plot.}

\item{width}{Plot width on the page.}

\item{height}{Plot height on the page.}
}
\description{
Used to build the UI for a corresponding call to \code{renderChart}.
}
