% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{renderChart}
\alias{renderChart}
\title{Chart Rendering}
\usage{
renderChart(expr, env = parent.frame(), quoted = F)
}
\arguments{
\item{expr}{An expression that generates a plot.}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression?}
}
\description{
Renders a reactive and interactve plot on an HTML canvas using Chart.js.
}
